+++ invenio_rdm_records/assets/semantic-ui/js/invenio_rdm_records/src/deposit/fields/FileUploader/FileUploader.js
@@ -1,5 +1,5 @@
 // This file is part of Invenio-RDM-Records
-// Copyright (C) 2020-2023 CERN.
+// Copyright (C) 2020-2025 CERN.
 // Copyright (C) 2020-2022 Northwestern University.
 // Copyright (C)      2022 Graz University of Technology.
 // Copyright (C)      2022 TU Wien.
@@ -45,7 +45,9 @@ export const FileUploaderComponent = ({
   ...uiProps
 }) => {
   // We extract the working copy of the draft stored as `values` in formik
-  const { values: formikDraft } = useFormikContext();
+  const { values: formikDraft, errors, initialErrors } = useFormikContext();
+  const hasError = (errors.files || initialErrors?.files) && files;
+
   const filesEnabled = _get(formikDraft, "files.enabled", false);
   const [warningMsg, setWarningMsg] = useState();
   const lockFileUploader = !isDraftRecord && filesLocked;
@@ -216,6 +218,7 @@ export const FileUploaderComponent = ({
       warningMsg={warningMsg}
       setWarningMsg={setWarningMsg}
       filesLocked={lockFileUploader}
+      hasError={hasError}
       {...uiProps}
     >
       <>
@@ -273,6 +276,7 @@ export const FileUploaderComponent = ({
             filesList={filesList}
             dropzoneParams={dropzoneParams}
             filesLocked={lockFileUploader}
+            hasError={hasError}
             filesEnabled={filesEnabled}
             deleteFile={deleteFile}
             decimalSizeDisplay={decimalSizeDisplay}
@@ -285,6 +289,7 @@ export const FileUploaderComponent = ({
                   filesList={filesList}
                   dropzoneParams={dropzoneParams}
                   filesLocked={lockFileUploader}
+                  hasError={hasError}
                   filesEnabled={filesEnabled}
                   deleteFile={deleteFile}
                   decimalSizeDisplay={decimalSizeDisplay}