+++ invenio_rdm_records/assets/semantic-ui/js/invenio_rdm_records/src/deposit/fields/FileUploader/FileUploaderArea.js
@@ -1,5 +1,5 @@
 // This file is part of Invenio-RDM-Records
-// Copyright (C) 2020-2024 CERN.
+// Copyright (C) 2020-2025 CERN.
 // Copyright (C) 2020-2022 Northwestern University.
 // Copyright (C) 2021-2022 Graz University of Technology.
 // Copyright (C)      2022 TU Wien.
@@ -202,6 +202,7 @@ const FileUploadBox = ({
   filesLocked,
   filesList,
   dragText,
+  hasError,
   uploadButtonIcon,
   uploadButtonText,
   openFileDialog,
@@ -225,7 +226,7 @@ const FileUploadBox = ({
           <Grid.Column mobile={16} tablet={7} computer={7}>
             <Button
               type="button"
-              primary
+              className={hasError ? "error" : "primary"}
               labelPosition="left"
               icon={uploadButtonIcon}
               content={uploadButtonText}
@@ -241,6 +242,7 @@ const FileUploadBox = ({
 FileUploadBox.propTypes = {
   filesLocked: PropTypes.bool.isRequired,
   filesList: PropTypes.array,
+  hasError: PropTypes.bool,
   dragText: PropTypes.string,
   uploadButtonIcon: PropTypes.node,
   uploadButtonText: PropTypes.string,
@@ -253,6 +255,7 @@ FileUploadBox.defaultProps = {
   uploadButtonIcon: undefined,
   uploadButtonText: undefined,
   openFileDialog: null,
+  hasError: false,
 };
 
 const FilesListTable = ({ filesLocked, filesList, deleteFile, decimalSizeDisplay }) => {