+++ invenio_rdm_records/requests/community_inclusion.py
@@ -107,7 +107,9 @@ class CommunityInclusion(RequestType):
     allowed_topic_ref_types = ["record"]
     needs_context = {
         "community_roles": ["owner", "manager", "curator"],
+        "record_permission": "preview",
     }
+    resolve_topic_needs = True
 
     available_actions = {
         "create": actions.CreateAction,