+++ tests/resources/test_resources.py
@@ -16,6 +16,7 @@ from io import BytesIO
 
 import arrow
 import pytest
+from invenio_accounts.testutils import login_user_via_session
 from invenio_requests import current_requests_service
 from marshmallow_utils.permissions import FieldPermissionError