+++ CHANGES.rst
@@ -1,6 +1,6 @@
 
 ..
-    Copyright (C) 2019-2024 CERN.
+    Copyright (C) 2019-2025 CERN.
     Copyright (C) 2019-2024 Northwestern University.
     Copyright (C) 2024      KTH Royal Institute of Technology.
     Copyright (C) 2024 Graz University of Technology.
@@ -11,6 +11,27 @@
 
 Changes
 =======
+
+Version v17.4.0 (released 2025-03-18)
+
+- deposit-ui: Support new error format with severity and description
+  - FeedbackLabel for creatibutors and license
+
+Version v17.3.0 (released 2025-03-11)
+
+- resources: add param to filter shared with my uploads
+  - returns record needs on entity resolution
+  - reindex associated request on parent access changes
+- service: change community submission actions on who can manage
+- links: add preview_html link
+
+Version v17.2.0 (released 2025-03-10)
+
+- views: signposting: files: fix filename encoding issues for downloads
+- resource_types: fix datapaper and interactiveresource datacite mapping
+- schema.org: add uploadDate for VideoObject serialization [+]
+- cff: add default "message" field
+- iiif: fix info request not being proxied
 
 Version v17.1.0 (released 2025-02-21)